Moved project plugins to correct package and added repos plugin
